+++ src/main/java/com/rh/folhaPagamento/model/FolhaDePagamento.java
@@ -1,14 +1,15 @@
 package com.rh.folhaPagamento.model;
+import java.math.BigDecimal;
 import java.util.List;
 
 public class FolhaDePagamento {
     private List<Funcionario> funcionarios;
-    private double totalBeneficios;
-    private double totalAdicionais;
-    private double totalDescontos;
-    private double salarioBruto;
-    private double salarioLiquido;
+    private BigDecimal totalBeneficios;
+    private BigDecimal totalAdicionais;
+    private BigDecimal totalDescontos;
+    private BigDecimal salarioBruto;
+    private BigDecimal salarioLiquido;
     private int mesReferencia;
     private int anoReferencia;
-    private long id;
+    private int id;
 }
